Create IntegerBinaryFunction and modularize its subclasses
